+++ src/collective/contact/widget/widgets.py
@@ -2,12 +2,12 @@ from z3c.form.interfaces import IFieldWidget
 import z3c.form.interfaces
 from z3c.form.widget import FieldWidget
 from zope.component import getUtility
-from zope.i18n import translate
+from zope.component.interfaces import ComponentLookupError
 from zope.interface import implementer, implements, Interface
 from zope.browserpage.viewpagetemplatefile import ViewPageTemplateFile
 from five import grok
-from Products.CMFPlone.utils import base_hasattr, safe_unicode
 
+from Products.CMFPlone.utils import base_hasattr, safe_unicode
 from plone.app.layout.viewlets.interfaces import IBelowContent
 from plone.app.layout.viewlets.interfaces import IHtmlHeadLinks
 from plone.formwidget.autocomplete.widget import (
@@ -15,108 +15,24 @@ from plone.formwidget.autocomplete.widget import (
     AutocompleteSelectionWidget)
 from plone.formwidget.autocomplete.widget import AutocompleteSearch as BaseAutocompleteSearch
 
-from . import _
-from .interfaces import (
+from collective.contact.widget import _
+from collective.contact.widget.interfaces import (
     IContactAutocompleteWidget,
     IContactAutocompleteSelectionWidget,
     IContactAutocompleteMultiSelectionWidget,
     IContactContent,
     IContactWidgetSettings,
     )
-from zope.component.interfaces import ComponentLookupError
+
 
 class PatchLoadInsideOverlay(grok.Viewlet):
     grok.context(Interface)
     grok.viewletmanager(IHtmlHeadLinks)
+    wait_msg = _(u"please wait")
+    tooltip_template = ViewPageTemplateFile('js/widget.js.pt')
 
     def render(self):
-        return """<script type="text/javascript">
-var ccw = {};
-$(document).ready(function() {
-  $(document).bind('formOverlayLoadSuccess', function(e, req, myform, api, pb, ajax_parent) {
-    ajax_parent.find('div').slice(0, 1).prepend(ajax_parent.find('.portalMessage').detach());
-  });
-  $(document).bind('loadInsideOverlay', function(e, el, responseText, errorText, api) {
-    var el = $(el);
-    var o = el.closest('.overlay-ajax');
-    var pbo = o.data('pbo');
-    var overlay_counter = parseInt(pbo.nt.substring(3, pbo.nt.length));
-    o.css({zIndex: 9998+overlay_counter});
-  });
-  ccw.fill_autocomplete = function (el, pbo, noform) {
-    var objpath = el.find('input[name=objpath]');
-    if (objpath.length) {
-        data = objpath.val().split('|');
-        var input_box = pbo.source.siblings('div').find('.querySelectSearch input');
-        formwidget_autocomplete_new_value(input_box, data[0], data[1]);
-        input_box.flushCache();
-        // trigger change event on newly added input element
-        var input = input_box.parents('.querySelectSearch').parent('div').siblings('.autocompleteInputWidget').find('input').last();
-        var url = data[3];
-        ccw.add_contact_preview(input, url);
-        input.trigger('change');
-    }
-    return noform;
-  };
-
-  var pendingCall = {timeStamp: null, procID: null};
-  ccw.add_contact_preview = function (input, url) {
-    if (url) {
-        input.siblings('.label')
-            .wrapInner('<a href="'+url+'" target="_new" class="link-tooltip">');
-    }
-  };
-
-  $(document).delegate('.link-tooltip', 'mouseleave', function() {
-    if (pendingCall.procID) {
-      clearTimeout(pendingCall.procID);
-      pendingCall.procID = null;
-    }
-  });
-  $(document).delegate('.link-tooltip', 'mouseenter', function() {
-    var trigger = $(this);
-    // don't open tooltip in tooltip
-    if (trigger.closest('.tooltip').length) {
-        return;
-    }
-    if (!trigger.data('tooltip')) {
-      if (pendingCall.procID) {
-        clearTimeout(pendingCall.procID);
-      }
-      var timeStamp = new Date();
-      var tooltipCall = function() {
-          var tip = $('<div class="tooltip pb-ajax" style="display:none">%s</div>')
-                .insertAfter(trigger);
-          trigger.tooltip({relative: true, position: "center right"});
-          var tooltip = trigger.tooltip();
-          tooltip.show();
-          var url = trigger.attr('href');
-          $.get(url, {ajax_load: new Date().getTime()}, function(data) {
-            tooltip.hide();
-            tooltip.getTip().html($('<div />').append(
-                    data.replace(/<script(.|\s)*?\/script>/gi, ""))
-                .find(common_content_filter));
-            if (pendingCall.timeStamp == timeStamp) {
-                tooltip.show();
-            }
-            pendingCall.procID = null;
-          });
-      }
-      pendingCall = {timeStamp: timeStamp,
-                     procID: setTimeout(tooltipCall, 500)};
-    }
-  });
-});
-</script>
-<style type="text/css">
-.tooltip {
-  overflow: hidden;
-}
-.tooltip, #calroot {
-  z-index: 99999;
-}
-</style>
-""" % translate(_(u"please wait"), context=self.request)
+        return self.tooltip_template()
 
 
 class TermViewlet(grok.Viewlet):
@@ -149,26 +65,6 @@ class TermViewlet(grok.Viewlet):
         return u"""<input type="hidden" name="objpath" value="%s" />""" % (
                     '|'.join([self.token, self.title, self.portal_type, self.url]))
 
-# Execute prepOverlay only if it hasn't been done yet, this avoid to have a
-# pbo undefined error when you have recursive overlays.
-OVERLAY_TEMPLATE = """
-$('#%(id)s-autocomplete').find('.%(klass)s').each(function() {
-    if ($(this).data('pbo') === undefined) {
-        $(this).prepOverlay({
-          subtype: 'ajax',
-          filter: common_content_filter+',#viewlet-below-content>*',
-          formselector: '%(formselector)s',
-          cssclass: 'overlay-contact-addnew',
-          closeselector: '%(closeselector)s',
-          noform: function(el, pbo) {return ccw.fill_autocomplete(el, pbo, 'close');},
-          config: {
-              closeOnClick: %(closeOnClick)s,
-              closeOnEsc: %(closeOnClick)s
-          }
-        });
-    }
-});
-"""
 
 class ContactBaseWidget(object):
     implements(IContactAutocompleteWidget)
@@ -192,6 +88,8 @@ function (event, data, formatted) {
     }(jQuery));
 }
 """
+    overlay_template = ViewPageTemplateFile('js/overlay.js.pt')
+    placeholder = _(u"Fill your search here...")
 
     @property
     def bound_source(self):
@@ -229,20 +127,18 @@ function (event, data, formatted) {
             else:
                 closeselector = action.get('closeselector',
                         '[name="form.buttons.cancel"]')
-                content += OVERLAY_TEMPLATE % dict(
-                        id=self.id,
+                content += self.overlay_template(**dict(
                         klass=action['klass'],
                         formselector=formselector,
                         closeselector=closeselector,
-                        closeOnClick=self.close_on_click and 'true' or 'false')
+                        closeOnClick=self.close_on_click and 'true' or 'false'))
 
         if include_default:
-            content += OVERLAY_TEMPLATE % dict(
-                    id=self.id,
+            content += self.overlay_template(**dict(
                     klass='addnew',
                     formselector='#form',
                     closeselector='[name="form.buttons.cancel"]',
-                    closeOnClick=self.close_on_click and 'true' or 'false')
+                    closeOnClick=self.close_on_click and 'true' or 'false'))
 
         return content